completed shopping cart add to cart

+++ INFT3050 project/INFT3050 project/Models/Stocktake/Stocktake.cs
@@ -1,4 +1,5 @@
 ﻿namespace INFT3050_project.Models.Stocktake
+
 {
     public class Stocktake
     {
@@ -6,7 +7,7 @@
         public int SourceId { get; set; }
         public int ProductId { get; set; }
         public int Quantity { get; set; }
-        public float Price { get; set; }
+        public double Price { get; set; }
         
     }
 }
